+++ grails-app/controllers/io/xh/hoist/PingController.groovy
@@ -10,16 +10,22 @@ import groovy.transform.CompileStatic
 import io.xh.hoist.security.AccessAll
 import io.xh.hoist.util.Utils
 
+
+/**
+ * A deliberately trivial endpoint. Provides a baseline check that a Hoist app server is up,
+ * running, and reachable without any additional expectations or overhead.
+ */
 @AccessAll
 @CompileStatic
 class PingController extends BaseController {
 
     def index() {
         renderJSON([
-                appCode  : Utils.appCode,
-                appName  : Utils.appName,
-                timestamp: System.currentTimeMillis(),
-                success  : true
+            appCode: Utils.appCode,
+            appName: Utils.appName,
+            timestamp: System.currentTimeMillis(),
+            success: true
         ])
     }
+
 }